refactor(runtime): one pipeline spine for both painters (#714)
The source→reducer glue was hand-mirrored line for line across
run_async (TUI) and floating::run — floating's own comment said
"mirrors runtime::driver::run_async". A per-line diff of the two
bodies found NINE identical steps (presence channel + exit watch,
build_source_set, event/scene/health channels, floor-caps atomics,
reducer_task spawn, SourceManager::spawn_with_health) and exactly
three real divergences, which STAY caller-side:

- boot_caps: the TUI measures the terminal (cap-clamped, headless
  arms), floating its window pixels — the documented over-seed sharp
  edge; the pipeline takes the seed as a parameter and never computes
  it.
- socket resolution and ConnectedSources: both values outlive boot
  (the Sources panel shows the path and mutates the live set).

runtime::pipeline::spawn_pipeline(...) -> Pipeline now owns the spine;
both callers keep only their preamble + the returned handles. The
spawns land on the ambient runtime (run_async's block_on, floating's
existing rt.enter guard — floating's explicit rt.spawn was equivalent
and is gone). Pipeline's _source_handles field must stay bound at the
call site (dropping it drops the source tasks) — said at the struct,
both callers bind it.

No behavior change intended; the module is codecov-excluded async glue
like its two callers (#103), so per the issue the verification is
dogfood, done on both entry points: `run --headless --projects-root
~/.claude/projects` against live CC sessions (the pipeline watched its
own dogfood command as a working sprite; clean SIGINT summary) and a
floating window boot (zero errors/panics in the log).

Closes #714
