+++ codecov.yml
@@ -69,6 +69,7 @@ ignore:
   - "crates/pixtuoid/src/sources_cli.rs"        # println presenters over the real config path (was main.rs; covered by the cli_json e2e binary run)
   - "crates/pixtuoid/src/tui/mod.rs"            # crossterm event loop + terminal lifecycle: reads/writes the real TTY
   - "crates/pixtuoid/src/runtime/driver.rs"     # async glue: tokio runtime + block_on + ctrl_c + socket bind (#103)
+  - "crates/pixtuoid/src/runtime/pipeline.rs"   # the shared source→reducer spine (#714): same channel/spawn glue class as driver.rs
   - "crates/pixtuoid/src/floating/mod.rs"       # floating run(): tokio runtime + winit EventLoop glue (the floating twin of driver.rs)
   - "crates/pixtuoid/src/floating/window.rs"    # winit ApplicationHandler: creates the real OS window + softbuffer surface
   # The testable floating logic IS counted: offscreen.rs (render seam) + geometry.rs (window/monitor rect math) + input.rs (the pure audio key-map).
